Red-black tree: batch process
This feature allows to construct a complicated tree from a piece of code
text instead of repeated manual typing of value and clicking of buttons.

The core idea is to define a call back function that remembers the late
iteration, and upon called, will fetch the code and execute it. It is then
passed into the animation manager, who, upon ending of a animation, will
try to execute the call back.

Additional notes:
1. A modal form is created to accept input code text. Currently
supported operations are insert, find, delete.
2. The animation is asynchronous, if we simply do animation in a
for-loop, things won't work out.
3. Js libs jquery-ui and jquery are referenced from /bower_components
dir; we shall add package management *.json file for them in later patches.